Use enums for Exif group and tag info structs

+++ src/swig-0_27_4/image_wrap.cxx
@@ -4534,7 +4534,7 @@ static PyObject* get_enum_typeobject(Exiv2::ByteOrder value) {
 
 
 static PyObject* py_from_enum(Exiv2::ByteOrder value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
@@ -4575,7 +4575,7 @@ static PyObject* get_enum_typeobject(Exiv2::AccessMode value) {
 
 
 static PyObject* py_from_enum(Exiv2::AccessMode value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
