+++ src/swig-0_28_0/error_wrap.cxx
@@ -4170,7 +4170,7 @@ static PyObject* get_enum_typeobject(Exiv2::LogMsg::Level value) {
 
 
 static PyObject* py_from_enum(Exiv2::LogMsg::Level value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
