+++ src/swig-0_28_0/value_wrap.cxx
@@ -5570,7 +5570,7 @@ static PyObject* get_enum_typeobject(Exiv2::TypeId value) {
 
 
 static PyObject* py_from_enum(Exiv2::TypeId value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
@@ -6098,7 +6098,7 @@ static PyObject* get_enum_typeobject(Exiv2::CommentValue::CharsetId value) {
 
 
 static PyObject* py_from_enum(Exiv2::CommentValue::CharsetId value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
@@ -6114,7 +6114,7 @@ static PyObject* py_from_enum(Exiv2::CommentValue::CharsetId value) {
 
 
 static PyObject* py_from_enum(Exiv2::ByteOrder value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
@@ -6157,7 +6157,7 @@ static PyObject* get_enum_typeobject(Exiv2::XmpValue::XmpArrayType value) {
 
 
 static PyObject* py_from_enum(Exiv2::XmpValue::XmpArrayType value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
@@ -6181,7 +6181,7 @@ static PyObject* get_enum_typeobject(Exiv2::XmpValue::XmpStruct value) {
 
 
 static PyObject* py_from_enum(Exiv2::XmpValue::XmpStruct value) {
-    PyObject* py_int = PyLong_FromLong(value);
+    PyObject* py_int = PyLong_FromLong(static_cast<long>(value));
     if (!py_int)
         return NULL;
     PyObject* result = PyObject_CallFunctionObjArgs(
